+++ ConsidSite/Models/Data/ICompaniesRepo.cs
@@ -5,7 +5,7 @@ namespace ConsidSite.Models.Data
 {
     public interface ICompaniesRepo
     {
-        Companies Create(string name, int organizationNumber, string notes);
+        Companies Create(string name, long organizationNumber, string notes);
         List<Companies> Read();
         Companies Read(Guid id);
         Companies Update(Companies company);
